Loop control is working
Stream ID is compared for all incoming openbridge connections.

The first incoming connection (system) that we see a stream on is used as the
source for that stream.

All other sources for that stream are ignored, for the duration of the stream.

This should enable multiple bridges carrying the same traffic, which gives us diverse
routing, fault tolerance and the ability to build a "mesh".

It also naturally uses the fastest route for that stream in preference to slower ones
as it implements a "first past the post" system, on a stream-by-stream basis. This means
it will naturally adapt to changing network conditions.

It is backwards-compatible with existing OBP as it makes no changes to the protocol at all.